adjust brightness of raw with scroll if in edit mode (trk)
Brightness adjustment used to depend only on rendering_raw (bool),
which could be unintuitive when in edit mode. To keep things simple
in browser caliban, scroll wheel will only be bound to adjustment of
the raw image, never the annotations. (If scroll wheel used to adjust
anything else in future, should be used with a modifier key.) So,
raw brightness isn't adjusted if we aren't looking at it (viewing annotations
only), but it is adjusted if we're rendering_raw or in edit_mode.